+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 
 # Change 5.2.5
+- Fixed Bluetooth auto sink switch 
+- Fixed app close sink change bug #24
 - Removed a lot of whitespaces from the files
 - Fixed float (again)
 - Removed all 'print' statements, and changed them into logs